clean up of minor issues
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@464 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ hazelcast/src/main/java/com/hazelcast/impl/ConcurrentMapManager.java
@@ -194,7 +194,7 @@ public final class ConcurrentMapManager extends BaseManager {
             setNewMembers:
             for (Address address : allAddress) {
                 Integer countInt = addressBlocks.get(address);
-                int count = (countInt == null) ? 0 : countInt.intValue();
+                int count = (countInt == null) ? 0 : countInt;
                 while (count < aveBlockOwnCount) {
                     if (lsBlocksToRedistribute.size() > 0) {
                         Block blockToMigrate = lsBlocksToRedistribute.remove(0);
